requirements.txt
zalo-python-sdk
dateparser
requests
flask
flask_cors
clarifai
scrapy
PIL
